# plan: cut unuseful search branch to improve performence.