If the prop is empty and selection is nil, we don't need to consider index scan.
If the prop is empty, we don't need to consider enforce sorter.
@shenli @coocood @zimulala @tiancaiamao @XuHuaiyu @winoros PTAL