@zimulala 2016-11-29
Fix CI @hanfei1991