@hanfei1991 2016-11-29
Before change:
BenchmarkBasic-4                      100000         15965 ns/op
BenchmarkTableScan-4                    2000        683003 ns/op
BenchmarkTableLookup-4                 20000         93759 ns/op
BenchmarkStringIndexScan-4              2000        700549 ns/op
BenchmarkStringIndexLookup-4           10000        105579 ns/op
BenchmarkIntegerIndexScan-4             2000        697651 ns/op
BenchmarkIntegerIndexLookup-4          10000        100078 ns/op
BenchmarkDecimalIndexScan-4             2000        752956 ns/op
BenchmarkDecimalIndexLookup-4          10000        108311 ns/op
After change:
BenchmarkBasic-4                      200000         16911 ns/op
BenchmarkTableScan-4                    2000        651861 ns/op
BenchmarkTableLookup-4                 20000         96550 ns/op
BenchmarkStringIndexScan-4              2000        695863 ns/op
BenchmarkStringIndexLookup-4           20000        102792 ns/op
BenchmarkIntegerIndexScan-4             2000        687823 ns/op
BenchmarkIntegerIndexLookup-4          10000        100590 ns/op
BenchmarkDecimalIndexScan-4             2000        763613 ns/op
BenchmarkDecimalIndexLookup-4          10000        104836 ns/op