@hanfei1991 2016-11-29
A more explicit bench result, before change:
BenchmarkExplainTableScan-4                30000         44607 ns/op
BenchmarkExplainTableLookup-4              20000         71500 ns/op
BenchmarkExplainStringIndexScan-4          20000         75814 ns/op
after change.
BenchmarkExplainTableScan-4                30000         41463 ns/op
BenchmarkExplainTableLookup-4              20000         70530 ns/op
BenchmarkExplainStringIndexScan-4          20000         74494 ns/op